support literal slot equal pair in ReplacePredicate
